Move safekeeper related stuff out of service.rs
